requirements_dev.txt
-r requirements.txt
-r requirements_client.txt

# Used in dev.py.
executor >= 21.0
click >=7.0,<8.0

# Documentation.
Sphinx
recommonmark

# Unit tests.
pytest ==6.0.1
pytest-cov == 2.10.0
pytest-redis==2.0.0
freezegun==0.3.15

# Optional dependencies.
redis==3.5.3
git+https://github.com/rq/rq.git@cf501518534429640a9a6d14e022f9329225e701#egg=rq

black ==19.10b0
flake8==3.8.3
flake8-import-order==0.18.1

# For CI.
coveralls ==2.1.2

python-dotenv==0.14.0
